#23
In fw, you can optionally supply a heuristic file in your project.

If not supplied, the gear checks the config "default_heuristic" for any 
string input. If the action is curate and there is no heuristic 
variable, the gear fails. Else, the string is then sent to "Curate", 
which will search the heudiconv module for the default heuristic as 
normal.